+++ Influx/Public/ConvertTo-InfluxLineString.ps1
@@ -96,15 +96,8 @@
                 if ($ExcludeEmptyMetric -and [string]::IsNullOrEmpty($MetricObject.Metrics[$Metric])) {
                     Write-Verbose "$Metric skipped as -ExcludeEmptyMetric was specified and the value is null or empty."
                 }
-                #if not a number wrap in "" and escape all influx special char
-                elseif ($MetricObject.Metrics[$Metric] -isnot [ValueType]) {
-                    $MetricValue = '"' + ($MetricObject.Metrics[$Metric] | Out-InfluxEscapeString -StringType FieldTextValue ) + '"'
-                    #Write output
-                    "$($Metric | Out-InfluxEscapeString)=$($MetricValue)"
-                }
-                #no need to escape numeric values
                 else {
-                    $MetricValue = $MetricObject.Metrics[$Metric]
+                    $MetricValue = $MetricObject.Metrics[$Metric] | Format-InfluxFieldValue
                     #Write output
                     "$($Metric | Out-InfluxEscapeString)=$($MetricValue)"
                 }