+++ Tests/Write-InfluxUDP.Tests.ps1
@@ -111,8 +111,8 @@ Describe "Write-InfluxUDP PS$PSVersion" {
             It 'Should call ConvertTo-UnixTimeNanosecond exactly 1 time' {
                 Should -Invoke ConvertTo-UnixTimeNanosecond -Exactly -Times 1 -Scope Context
             }
-            It 'Should call Out-InfluxEscapeString exactly 8 times' {
-                Should -Invoke Out-InfluxEscapeString -Exactly -Times 8 -Scope Context
+            It 'Should call Out-InfluxEscapeString exactly 6 times' {
+                Should -Invoke Out-InfluxEscapeString -Exactly -Times 6 -Scope Context
             }
             It 'Should call Invoke-UDPSendMethod exactly 1 time' {
                 Should -Invoke Invoke-UDPSendMethod -Exactly -Times 1 -Scope Context